+++ db/migrate/20180826140638_add_interval_to_tasks.rb
@@ -1,0 +1,6 @@
+class AddIntervalToTasks < ActiveRecord::Migration[5.2]
+  def change
+    add_column :tasks, :interval_type, :integer, null: true
+    add_column :tasks, :interval_value, :integer, null: true
+  end
+end