+++ test/controllers/tasks_controller_test.rb
@@ -2,6 +2,7 @@ require 'test_helper'
 
 class TasksControllerTest < ActionDispatch::IntegrationTest
   setup do
+    login_user(users(:one))
     @task = tasks(:one)
   end
 
@@ -33,9 +34,23 @@ class TasksControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
-  test "should update task" do
-    patch task_url(@task), params: { task: { done: @task.done, due_date: @task.due_date, title: @task.title } }
-    assert_redirected_to task_url(@task)
+  context "update" do
+    should "update task" do
+      patch task_url(@task), params: { task: { done: @task.done, due_date: @task.due_date, title: @task.title } }
+      assert_redirected_to task_url(@task)
+    end
+
+    should "include next task if any" do
+      task = FactoryBot.create(:task, interval_type: 'n_days_after', interval_value: 1)
+      patch task_url(task, format: :json), params: {task: {done: true}}
+      assert_equal false, JSON.parse(response.body)["next_task"]["done"]
+    end
+
+    should "next task is nil if none" do
+      task = FactoryBot.create(:task)
+      patch task_url(task, format: :json), params: {task: {done: true}}
+      assert_nil JSON.parse(response.body)["next_task"]
+    end
   end
 
   test "should destroy task" do