+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# (not yet)
+# HEAD
 
+- feat: Recurring tasks
 - fix: project_id is set to 0 when '---' is selected
 
 # v0.2.6 (2018-08-27)